remove duplicate classes
