+++ pitest/src/main/java/org/pitest/classinfo/ClassPointer.java
@@ -35,18 +35,4 @@ class DefaultClassPointer implements ClassPointer {
 
 }
 
-class DeferredClassPointer implements ClassPointer {
-  private final Repository repository;
-  private final ClassName  name;
 
-  DeferredClassPointer(final Repository repository, final ClassName name) {
-    this.repository = repository;
-    this.name = name;
-  }
-
-  @Override
-  public Optional<ClassInfo> fetch() {
-    return this.repository.fetchClass(this.name);
-  }
-
-}
