--- pitest/src/test/java/org/pitest/classinfo/NameToClassInfoTest.java
@@ -1,18 +1,0 @@
-package org.pitest.classinfo;
-
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-
-import org.junit.Test;
-
-public class NameToClassInfoTest {
-
-  @Test
-  public void shouldFetchClassWhenApplied() {
-    final Repository repository = mock(Repository.class);
-    final NameToClassInfo testee = new NameToClassInfo(repository);
-    testee.apply(ClassName.fromString("foo"));
-    verify(repository).fetchClass(ClassName.fromString("foo"));
-  }
-
-}
